chore(widget): rename stores to store
