+++ src/widget/stores/index.ts
