+++ src/widget/stores/useChartStore.ts
